YahBoom GameHandle MicroBit Code was made
The YahBoom Game Handle is a nice little controller built for the Micro:Bit with zero explanation on how to use the hardware without their pre-built hex image files. While this code does not make use of the motor for vibration feedback nor the piezo buzzer, it does make use of all buttons and the joystick.